\name{tranSMART-package}
\alias{tranSMART-package}
\alias{tranSMART}
\docType{package}
\title{
Functionality for interacting with the tranSMART data warehouse.
}
\description{
Functions that prepare SQL queries that run against the tranSMART data warehouse. Some extra functionality exists to transform the data after querying and before rendering to user.
}
\details{
\tabular{ll}{
Package: \tab tranSMART\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-13\cr
}
}
\author{
mmcduffie

Maintainer: mmcduffie <mmcduffie@recomdata.com>
}
\references{
\tab
}
\tab
\keyword{ package }
\seealso{
\tab
}
\examples{

ls("package:tranSMART")

lsf.str("package:tranSMART")

}
